requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.1
Django
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
frozenlist==1.5.0
idna==3.10
msgpack==1.1.0
multidict==6.1.0
pillow==11.1.0
propcache==0.2.1
PyJWT==2.10.1
redis==5.2.1
reportlab==4.3.0
requests==2.32.3
sqlparse==0.5.3
stripe==11.5.0
twilio==9.4.4
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
yarl==1.18.3
